work in progess: conversion to right rec grammer
